Makefile: Better method to store build flags
Before, the .build_infos file was included (via '-include') directly in
the Makefile at the start. Now, we sed into .build_infos to get the right flags.
This is overall a cleaner way of doing state saving, as it does not rely on putting Makefile variable names in other files.