Buffer: moved to a handle/reference system
Now, a Buffer doesn't hold directly a VkBuffer (and associated
VkMemory), it references it. A VkBuffer is only created when calling
explicitly Buffer::create_*_buffer().
For example:

Buffer a = create_vertex_buffer(...);
Buffer b = a; -> b points to the same VkBuffer as a, there is no
allocation here.
The new VkBuffer will be freed only when a AND b's destructor gets
called.